requirements.txt
aws-xray-sdk==2.8.0
boto3==1.18.53
botocore==1.21.53
Django==3.2.7
django-environ==0.7.0
django-cors-headers==3.9.0
djangorestframework==3.12.4
pymysql==1.0.2
python-dateutil==2.8.2
google-auth==2.2.1
Werkzeug==2.0.1
libica==1.1.0
libumccr==0.1.0
gspread==4.0.1
gspread-pandas==2.3.0
sample-sheet==0.12.0
requests==2.26.0
pandas==1.3.3
cachetools==4.2.4
drf_yasg==1.20.0
